Se implementan las modificaciones de la entidad máquina en la capa de servicios, dto y data y se crean los dtos de periodo y parámetro

+++ Tier.Dto/Asesor.cs
@@ -13,7 +13,7 @@ namespace Tier.Dto
         public string nombre { get; set; }
 
         [Column(Name = "comision")]
-        public Nullable<decimal> comision { get; set; }
+        public Nullable<Single> comision { get; set; }
 
         [Column(Name = "telefono")]
         public string telefono { get; set; }
